# docs/swagger.yaml
openapi: 3.1.1
info:
  title: Open Identity Certification with OpenID Connect (OIDC²)
  description: Authorization Server middleware for requesting Identity Certification Tokens (ICT).
  license:
    name: GNU General Public License v3.0
    url: https://github.com/JonasPrimbs/oidc-e2ea-server/blob/main/LICENSE
  contact: 
    name: Jonas Primbs
    email: mail@jonasprimbs.de
  version: 0.2.0
servers:
  - description: Publicly accessible API endpoint for testing with the OIDC E2E Test Environment.
    url: https://op.oidc-e2e.primbs.dev/realms/koala/protocol/openid-connect/token
  - description: Local development API endpoint for testing with a local OpenID Provider (e.g., Keycloak).
    url: http://op.localhost/realms/koala/protocol/openid-connect/token
paths:
  /:
    summary: The token endpoint.
    description: |
      This is the endpoint for requesting tokens from the authorization server.
      The application proxies all requests to the authorization server except token exchange requests for an Identity Certification Token, which are handled by the application itself.
    post:
      summary: Request a token.
      description: |
        This method proxies all token requests to the authorization server except token exchange requests for an Identity Certification Token.
        If an Identity Certification Token is requested, the application validates the request and generates the token itself if the request is valid.
      operationId: tokenRequest
      security:
        - {}
        - client_basic_auth: []
      requestBody:
        description: OAuth 2.0 Token Request body compliant with RFC 6749, RFC 8693, RFC 9449, and OpenID Connect Core 1.0
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - grant_type
                - client_id
              properties:
                grant_type:
                  description: REQUIRED. The grant type being used.
                  type: string
                  enum:
                    - authorization_code
                    - password
                    - client_credentials
                    - refresh_token
                    - urn:ietf:params:oauth:grant-type:device_code
                    - urn:ietf:params:oauth:grant-type:jwt-bearer
                    - urn:ietf:params:oauth:grant-type:saml2-bearer
                    - urn:ietf:params:oauth:grant-type:ciba
                    - urn:ietf:params:oauth:grant-type:token-exchange
                client_id:
                  description: REQUIRED. Client identifier.
                  type: string
                  pattern: ^[!-~]+$
                code:
                  description: REQUIRED for authorization_code grant.
                  type: string
                  pattern: ^[!-~]+$
                redirect_uri:
                  description: REQUIRED if included in authorization request.
                  type: string
                  pattern: ^[a-z][a-z0-9+.-]*:\/\/([^#\s]+)$
                client_secret:
                  description: Required for confidential clients.
                  type: string
                  pattern: ^[!-~]+$
                refresh_token:
                  description: REQUIRED for refresh_token grant.
                  type: string
                  pattern: ^[a-zA-Z0-9._~\-\/+=]+$
                scope:
                  description: OPTIONAL. Space-delimited scope values.
                  type: string
                  pattern: ^[!#-\[\]-~]+$
                resource:
                  description: OPTIONAL (RFC 8693). Target API resource.
                  type: string
                  pattern: ^[!-~]+$
                audience:
                  description: OPTIONAL. Intended audience for the token.
                  type: string
                  pattern: ^[a-z][a-z0-9+.-]*:\/\/([^#\s]+)$
                subject_token:
                  description: OPTIONAL (RFC 8693). Subject token for token exchange.
                  type: string
                  pattern: ^[a-zA-Z0-9._~\-\/+=]+$
                subject_token_type:
                  description: OPTIONAL (RFC 8693). Subject token type identifier.
                  type: string
                  pattern: ^urn:[a-z0-9][a-z0-9-]{0,31}:[a-zA-Z0-9()+,\-.:=@;$_!*'%/?#]+$
                actor_token:
                  description: OPTIONAL (RFC 8693). Actor token for delegation.
                  type: string
                  pattern: ^[a-zA-Z0-9._~\-\/+=]+$
                actor_token_type:
                  description: OPTIONAL (RFC 8693). Actor token type identifier.
                  type: string
                  pattern: ^urn:[a-z0-9][a-z0-9-]{0,31}:[a-zA-Z0-9()+,\-.:=@;$_!*'%/?#]+$
                dpop:
                  description: OPTIONAL (RFC 9449). DPoP proof JWT.
                  type: string
                  pattern: ^[a-zA-Z0-9\-_]+\.[a-zA-Z0-9\-_]+\.[a-zA-Z0-9\-_]+$
      responses:
        "200":
          description: Token successfully issued
          headers:
            Cache-Control:
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                type: object
                required:
                  - access_token
                  - token_type
                properties:
                  access_token:
                    description: The access token issued by the authorization server.
                    type: string
                    pattern: ^[a-zA-Z0-9._~\-\/+=]+$
                  token_type:
                    description: Token type (typically "Bearer" or "DPoP").
                    type: string
                    enum:
                      - Bearer
                      - DPoP
                      - N_A
                  expires_in:
                    description: Lifetime in seconds of the access token.
                    type: integer
                    format: int64
                    example: 3600
                  refresh_token:
                    description: Refresh token for obtaining new access tokens.
                    type: string
                    pattern: ^[a-zA-Z0-9._~\-\/+=]+$
                  scope:
                    description: Scope of the access token.
                    type: string
                    pattern: ^[!#-\[\]-~]+$
                  id_token:
                    description: OPTIONAL (OpenID Connect). ID Token (JWT).
                    type: string
                    pattern: ^[a-zA-Z0-9\-_]+\.[a-zA-Z0-9\-_]+\.[a-zA-Z0-9\-_]+$
                  dpop_jkt:
                    description: OPTIONAL (RFC 9449). DPoP proof key thumbprint.
                    type: string
                    pattern: ^[a-zA-Z0-9\-_]{43}$
                  c_hash:
                    description: OPTIONAL (OpenID Connect). Code hash for authorization code flow.
                    type: string
                    pattern: ^([a-zA-Z0-9\-_]{22}|[a-zA-Z0-9\-_]{32}|[a-zA-Z0-9\-_]{43})$
                  issued_token_type:
                    description: OPTIONAL (RFC 8693). The type of the issued token.
                    type: string
                    pattern: ^urn:ietf:params:oauth:token-type:[a-zA-Z0-9._\-]+$
                  requested_token_type:
                    description: OPTIONAL (RFC 8693). The type of the requested token for confirmation.
                    type: string
                    pattern: ^urn:ietf:params:oauth:token-type:[a-zA-Z0-9._\-]+$
                  state:
                    description: OPTIONAL. State value from the authorization request.
                    type: string
                    pattern: ^[a-zA-Z0-9._~\-]{8,512}$
                  client_id:
                    description: OPTIONAL. Client identifier for which the token was issued for confirmation.
                    type: string
                    pattern: ^[!-~]+$
                  not-before-policy:
                    description: OPTIONAL (proprietary for Keycloak). The time before which the token MUST NOT be accepted for processing.
                    type: integer
                    format: int64
        "400":
          description: Invalid request or invalid grant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseError'
        "401":
          description: Client authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseError'
        "403":
          description: Client unauthorized to use this grant type or insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseError'
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStatus'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStatus'
    options:
      summary: CORS preflight request
      description: |
        This method is used for CORS preflight requests. It responds with appropriate CORS headers to allow cross-origin requests from authorized origins.
      operationId: tokenRequestPreflight
      responses:
        "204":
          description: No content, CORS preflight successful
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              example: "https://client.example.com"
            Access-Control-Allow-Methods:
              schema:
                type: string
              example: "POST, OPTIONS"
            Access-Control-Allow-Headers:
              schema:
                type: string
              example: "Content-Type, Authorization"
            Access-Control-Max-Age:
              schema:
                type: integer
                format: int64
              example: 86400
components:
  schemas:
    TokenResponseError:
      description: Error status response compliant with RFC 6749 and OpenID Connect Core 1.0
      type: object
      required:
        - error
      properties:
        error:
          description: Error code as defined in RFC 6749 and OpenID Connect Core 1.0
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - invalid_target
            - unsupported_token_type
            - authorization_pending
            - slow_down
            - access_denied
            - expired_token
            - use_dpop_nonce
            - invalid_dpop_proof
            - interaction_required
            - login_required
            - account_selection_required
        error_description:
          description: Human-readable ASCII encoded text providing additional information about the error
          type: string
        error_uri:
          description: URI identifying a human-readable web page with information about the error
          type: string
          format: uri
    ErrorStatus:
      description: Generic error status response
      type: object
      required:
        - error
      properties:
        error:
          description: Error code identifying the error condition
          type: string
          enum:
            - method_not_allowed
            - too_many_requests
            - server_error
        error_description:
          description: Human-readable ASCII encoded text providing additional information about the error
          type: string
  securitySchemes:
    client_basic_auth:
      description: HTTP Basic Authentication for client authentication
      type: http
      scheme: basic